--- Explorer/Assets/Plugins/CommandTerminal/ConsoleBuffer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 12f1ab2324bd74537b3fed066a4291fd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
